Finished get players, working on getting puuids.
get_top_players now correctly pulls all the players from the divisions in EU1, NA1, and KR and stores them in a dataframe.

The next step is to pull in the puuid by calling the helper function get_puuid which takes in a summonerId and a region and returns the puuid.